refactor: remove unused APPLICATION_PROXY_METHODS and simplify proxy route methods refactor: update test client import to use FastAPI's TestClient directly refactor: streamline migration error handling by inlining exception chain logic refactor: change get_pages to a synchronous function for improved performance refactor: replace synchronous migration function with an inline command upgrade call refactor: remove deprecated testing module and update documentation references refactor: delete unused SVG components to clean up the codebase
