+++ sdk/longlink/.static/new/tests/test_app.py
@@ -1,4 +1,4 @@
-from longlink.testing import TestClient
+from fastapi.testclient import TestClient
 
 
 def test_healthcheck_returns_ok_payload() -> None:
