--- sdk/longlink/testing.py
@@ -1,5 +1,0 @@
-from fastapi.testclient import TestClient as FastAPITestClient
-
-
-class TestClient(FastAPITestClient):
-    """LongLink test client facade over FastAPI's test client."""
